expense calculator modification

+++ python_introduction/finance_calculator.py
@@ -5,6 +5,6 @@ monthly_savings = float(monthly_income) - float(monthly_expense);
 
 annual_interest_rate = 0.05;
 
-Projected_savings = monthly_savings * 12 + float(monthly_savings*12*0.05);
+Projected_savings = float(monthly_savings) * 12 + float(monthly_savings*12*0.05);
 
 print(f"Your projected savings for the year is: {Projected_savings:.2f}");
